m8: Add Quality of Service Manager (QoSMgr)
The idea behind QoSMgr is to have a centralized place in the system to setup
quality of service rules that work outside the cpufreq governor. Wrote a parser
for a small language specifying these rules as boolean expressions and assignments
over sysfs nodes. The design also constantly monitors the system for the change in
value of boolean filters specified in the rules, re-evaluates the filters whose
value has changed, and then programs the rules for the filters that turn from false
to true.